scripts/claude_sheets_example.py
#!/usr/bin/env python3
"""
Example: Claude + Google Sheets automation
Requires: pip install google-api-python-client anthropic
"""

import anthropic
from google.oauth2.credentials import Credentials
from googleapiclient.discovery import build
import json

class ClaudeSheets:
    def __init__(self, anthropic_api_key, google_creds_path):
        self.claude = anthropic.Anthropic(api_key=anthropic_api_key)
        # You'd need Google OAuth2 setup here
        # self.sheets = build('sheets', 'v4', credentials=creds)
        
    def analyze_and_update_sheet(self, sheet_id, range_name, task_description):
        """
        1. Read data from Google Sheet
        2. Send to Claude for analysis/processing
        3. Update sheet with Claude's response
        """
        
        # Step 1: Read sheet data
        # result = self.sheets.spreadsheets().values().get(
        #     spreadsheetId=sheet_id, range=range_name).execute()
        # data = result.get('values', [])
        
        # Step 2: Ask Claude to process the data
        prompt = f"""
        I have this data from a Google Sheet: {data}
        
        Task: {task_description}
        
        Please analyze this data and provide:
        1. Your analysis
        2. Any calculations needed
        3. New data to add to the sheet (in the same format)
        
        Return as JSON with keys: analysis, calculations, new_data
        """
        
        response = self.claude.messages.create(
            model="claude-3-5-sonnet-20241022",
            max_tokens=1000,
            messages=[{"role": "user", "content": prompt}]
        )
        
        # Step 3: Parse Claude's response and update sheet
        claude_result = json.loads(response.content[0].text)
        
        # Update the sheet with new data
        # self.sheets.spreadsheets().values().update(
        #     spreadsheetId=sheet_id,
        #     range=range_name,
        #     valueInputOption='RAW',
        #     body={'values': claude_result['new_data']}
        # ).execute()
        
        return claude_result

# Example usage:
# cs = ClaudeSheets('your-anthropic-key', 'google-creds.json')
# result = cs.analyze_and_update_sheet('sheet-id', 'A1:C10', 'Calculate totals and add trends')