requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
jinja2>=3.1.2
python-multipart>=0.0.6
websockets>=12.0
anthropic>=0.3.0
google-api-python-client>=2.0.0
google-auth-oauthlib>=1.0.0
google-auth>=2.0.0
pandas>=2.0.0
python-dotenv>=1.0.0
httpx>=0.25.0
